Decompose analyze/pipeline.rs monolith into cohesive modules
The prior lib.rs/mod.rs convention pass relocated the 3.8k-line analyze
body wholesale into pipeline.rs. Split it along its natural seams so each
file owns one concern, leaving pipeline.rs as the orchestration core plus
its end-to-end tests and analyze/mod.rs as docs + mod decls + re-exports:

- selection.rs  Selection parameters + builders
- facets.rs     facet resolution
- window.rs     time-window / mode parsing
- history.rs    git-topology history resolution
- load.rs       concurrent object load + fold
- dataset.rs    dataset selection orchestration

Cross-file callers bump the shared helpers/fields they reach to
pub(crate); co-located private-helper unit tests move with their code.
