+++ packages/cargo-bench-history/src/analyze/mod.rs
@@ -13,17 +13,27 @@
 //! storage- and git-generic orchestrator the in-memory tests drive.
 
 pub(crate) mod bless;
+mod dataset;
 pub(crate) mod examine;
+mod facets;
+mod history;
 pub(crate) mod list;
+mod load;
 mod pipeline;
 pub(crate) mod prune;
+mod selection;
+mod window;
 
 pub(crate) use cargo_bench_history_core::analyze::{
     ReportFormat, Series, SeriesFilter, StorageKey, apply_blessings, format_value,
 };
-pub(crate) use pipeline::{
-    AutoFacets, DirtyTipPolicy, ResolvedHistory, RunIndex, Selection, WindowEdge,
-    detect_auto_facets, dirty_base_exception_warning, empty_history_hint, execute,
-    facet_filtered_candidates, parse_since, parse_until, resolve_base_name, resolve_base_ref,
-    resolve_facets, resolve_history, resolve_now, select_dataset, window_excludes,
+pub(crate) use dataset::{empty_history_hint, select_dataset};
+pub(crate) use facets::{AutoFacets, resolve_facets};
+pub(crate) use history::{
+    DirtyTipPolicy, ResolvedHistory, dirty_base_exception_warning, resolve_base_name,
+    resolve_base_ref, resolve_history,
 };
+pub(crate) use load::{RunIndex, facet_filtered_candidates};
+pub(crate) use pipeline::{detect_auto_facets, execute, resolve_now};
+pub(crate) use selection::Selection;
+pub(crate) use window::{WindowEdge, parse_since, parse_until, window_excludes};
