+++ packages/cargo-bench-history/src/analyze/pipeline.rs
@@ -1,9 +1,9 @@
-//! Implementation of the `analyze` pipeline: the orchestration entry points
-//! ([`execute`], [`analyze_with`]) and the pure selection, window-resolution,
-//! parsing, and reporting logic they compose. The parent module re-exports the
-//! surface the sibling query commands (`list`, `prune`, `examine`, `bless`) share.
+//! The `analyze` orchestration entry points: [`execute`] wires the real adapters
+//! and [`analyze_with`] is the storage- and git-generic orchestrator that the sibling
+//! modules (`selection`, `facets`, `load`, `dataset`, `history`, `window`) compose.
+//! The parent module re-exports the surface the sibling query commands
+//! (`list`, `prune`, `examine`, `bless`) share.
 
-use std::collections::{BTreeMap, HashMap};
 use std::io::IsTerminal;
 use std::path::Path;
 use std::sync::Arc;
@@ -11,39 +11,31 @@ use std::time::Instant;
 
 use anyspawn::Spawner;
 use cargo_bench_history_core::analyze::{
-    AnalysisConfig, AnalysisContext, AnalysisMode, BlessingPlacement, DEFAULT_SUMMARY_LIMIT,
-    DiscriminantSetQuery, FacetFilter, ReportInput, RunPoints, Series, SeriesBuilder, SeriesFilter,
-    SetSummary, StorageKey, apply_blessings, balanced_chunk_sizes, find_changes_spawned, parse_key,
-    render, render_markdown_summary, select_commits, worker_count,
+    AnalysisConfig, AnalysisContext, DEFAULT_SUMMARY_LIMIT, ReportInput, Series, SeriesFilter,
+    SetSummary, apply_blessings, find_changes_spawned, render, render_markdown_summary,
 };
-use futures::{StreamExt as _, TryStreamExt as _};
-use jiff::civil::Date;
-use jiff::tz::TimeZone;
-use jiff::{Span, Timestamp};
-use nonempty::NonEmpty;
+use jiff::Timestamp;
 use tick::Clock;
 
+use super::dataset::{empty_history_hint, select_dataset};
+use super::facets::AutoFacets;
+use super::history::dirty_base_exception_warning;
+use super::selection::Selection;
 use crate::config::{Config, load_config};
 use crate::git_history::{GitHistory, SystemGitHistory};
 use crate::machine::resolve_machine_key;
-use crate::model::{
-    BenchmarkIdPrefix, BlessingRecord, DiscriminantSet, Engine, STORAGE_VERSION, sanitize_segment,
-};
+use crate::model::DiscriminantSet;
 use crate::output::{
     OutputSelection, OutputWriter, TokioOutputWriter, emit, emit_markdown_summary,
 };
 use crate::probe::{EnvironmentProbe, SystemProbe};
 use crate::report::{Reporter, ReporterExt, StderrReporter};
-use crate::storage::{Storage, StorageFacade, project_objects_prefix, resolve_storage};
-use crate::text::count_noun;
+use crate::storage::{Storage, StorageFacade, resolve_storage};
 use crate::wiring::{
     cache_env, resolve_cache_path, resolve_config_path, resolve_local_path, resolve_project_id,
     resolve_repo, storage_env,
 };
-use crate::{
-    AnalyzeOptions, BlessOptions, ExamineOptions, ListOptions, PruneOptions, RunError, RunOutcome,
-    UnblessOptions,
-};
+use crate::{AnalyzeOptions, RunError, RunOutcome};
 
 /// The real `analyze`: load configuration, wire the configured storage and git
 /// history, and orchestrate.
@@ -303,1544 +295,6 @@ where
     })
 }
 
-/// The data-set selection parameters shared by the query commands: which stored
-/// objects to consider (facets + `--since` / `--until`) and how to resolve the git
-/// timeline (`--repo` is resolved by the caller into the [`GitHistory`] adapter;
-/// `--context` / `--base` / `--no-dirty` steer the topology query). Analyze's
-/// benchmark-prefix scope is deliberately *not* here: it filters which series are
-/// built, not which runs load.
-///
-/// Each facet (`engine` / `target_triple` / `machine_key`) carries the raw,
-/// repeatable command-line values; [`resolve_facets`] turns them into
-/// [`FacetFilter`]s, applying the current-machine auto-detect default and the
-/// `all` keyword.
-pub(crate) struct Selection<'a> {
-    context: Option<&'a str>,
-    pub(crate) base: Option<&'a str>,
-    no_dirty: bool,
-    since: Option<&'a str>,
-    until: Option<&'a str>,
-    engine: &'a [String],
-    target_triple: &'a [String],
-    machine_key: &'a [String],
-    /// Explicit `--mode` override; `None` lets the mode auto-detect from topology.
-    mode_override: Option<AnalysisMode>,
-}
-
-impl<'a> Selection<'a> {
-    fn from_analyze(options: &'a AnalyzeOptions) -> Result<Self, RunError> {
-        Ok(Self {
-            context: options.context.as_deref(),
-            base: options.base.as_deref(),
-            no_dirty: options.no_dirty,
-            since: options.since.as_deref(),
-            until: options.until.as_deref(),
-            engine: &options.engine,
-            target_triple: &options.target_triple,
-            machine_key: &options.machine_key,
-            mode_override: parse_mode(options.mode.as_deref())?,
-        })
-    }
-
-    pub(crate) fn from_list(options: &'a ListOptions) -> Self {
-        Self {
-            context: options.context.as_deref(),
-            base: options.base.as_deref(),
-            no_dirty: options.no_dirty,
-            since: options.since.as_deref(),
-            until: options.until.as_deref(),
-            engine: &options.engine,
-            target_triple: &options.target_triple,
-            machine_key: &options.machine_key,
-            mode_override: None,
-        }
-    }
-
-    pub(crate) fn from_examine(options: &'a ExamineOptions) -> Self {
-        Self {
-            context: options.context.as_deref(),
-            base: options.base.as_deref(),
-            no_dirty: options.no_dirty,
-            since: options.since.as_deref(),
-            until: options.until.as_deref(),
-            engine: &options.engine,
-            target_triple: &options.target_triple,
-            machine_key: &options.machine_key,
-            mode_override: None,
-        }
-    }
-
-    pub(crate) fn from_prune(options: &'a PruneOptions) -> Self {
-        Self {
-            context: options.context.as_deref(),
-            base: options.base.as_deref(),
-            // `prune` resolves the data set with dirty admission always on; the
-            // base-tip exception is applied unconditionally (see
-            // `DirtyTipPolicy::Always`), and the per-object scope (`--dirty` /
-            // `--clean`) decides which runs are actually removed.
-            no_dirty: false,
-            since: options.since.as_deref(),
-            until: options.until.as_deref(),
-            engine: &options.engine,
-            target_triple: &options.target_triple,
-            machine_key: &options.machine_key,
-            mode_override: None,
-        }
-    }
-
-    /// Selection facets for `bless`. Only the discriminant facets (and `base`)
-    /// matter: a blessing always acts at the current commit, so it has no
-    /// `context` / `since` / topology selectors.
-    pub(crate) fn from_bless(options: &'a BlessOptions) -> Self {
-        Self {
-            context: None,
-            base: options.base.as_deref(),
-            no_dirty: false,
-            since: None,
-            until: None,
-            engine: &options.engine,
-            target_triple: &options.target_triple,
-            machine_key: &options.machine_key,
-            mode_override: None,
-        }
-    }
-
-    /// Selection facets for `unbless`. Mirrors [`from_bless`](Self::from_bless).
-    pub(crate) fn from_unbless(options: &'a UnblessOptions) -> Self {
-        Self {
-            context: None,
-            base: options.base.as_deref(),
-            no_dirty: false,
-            since: None,
-            until: None,
-            engine: &options.engine,
-            target_triple: &options.target_triple,
-            machine_key: &options.machine_key,
-            mode_override: None,
-        }
-    }
-}
-
-/// The current machine's auto-detected facet values, used as the default when a
-/// query facet is omitted (see the *Discriminant set & query facets* section of
-/// `DESIGN.md`).
-///
-/// Production probes these once (host triple from `rustc -vV`, machine key from
-/// the hardware fingerprint); tests pass deterministic literals. There is no auto
-/// engine — a bare query analyzes every engine — so only the triple and machine
-/// key are detected.
-#[derive(Clone, Debug)]
-pub(crate) struct AutoFacets {
-    /// The host target triple (`rustc -vV` host).
-    pub(crate) triple: String,
-    /// The host machine fingerprint.
-    pub(crate) machine_key: String,
-}
-
-/// One commit's run tally within a discriminant set, the granularity the report
-/// summaries and the `list runs` breakdown need.
-#[derive(Clone, Debug)]
-pub(crate) struct CommitCounts {
-    /// The commit the runs were measured against (full commit ID, or a label in tests).
-    pub(crate) commit: String,
-    /// Clean (committed-tree) runs recorded on the commit.
-    pub(crate) clean: usize,
-    /// Dirty (uncommitted-tree) snapshots recorded on the commit.
-    pub(crate) dirty: usize,
-}
-
-/// Compact per-set, per-commit run tallies kept *in place of* a retained copy of
-/// every loaded object.
-///
-/// A long history holds tens of thousands of run objects, each carrying every
-/// benchmark; keeping them all resident alongside the reconstructed series is what
-/// drove analysis into tens of gigabytes. The analysis only needs the series plus
-/// these aggregate counts (total runs, per-set runs, and the per-commit breakdown
-/// the listing renders), so each parsed run is folded into the series and dropped,
-/// updating this index as it goes.
-#[derive(Clone, Debug, Default)]
-pub(crate) struct RunIndex {
-    total: usize,
-    sets: BTreeMap<DiscriminantSet, BTreeMap<usize, CommitCounts>>,
-}
-
-impl RunIndex {
-    /// An empty index.
-    fn new() -> Self {
-        Self::default()
-    }
-
-    /// Records one run on `commit` (first-parent position `topo_index`) in `set`.
-    fn record(&mut self, set: &DiscriminantSet, topo_index: usize, commit: &str, dirty: bool) {
-        self.total = self.total.saturating_add(1);
-        let entry = self
-            .sets
-            .entry(set.clone())
-            .or_default()
-            .entry(topo_index)
-            .or_insert_with(|| CommitCounts {
-                commit: commit.to_owned(),
-                clean: 0,
-                dirty: 0,
-            });
-        if dirty {
-            entry.dirty = entry.dirty.saturating_add(1);
-        } else {
-            entry.clean = entry.clean.saturating_add(1);
-        }
-    }
-
-    /// Folds another index's tallies into this one.
-    ///
-    /// The recombination step of the parallel fold: each worker records the runs of
-    /// its disjoint object chunk into its own index, and the main thread merges the
-    /// per-worker indices. Because the chunks are disjoint, summing the totals and
-    /// the per-`(set, commit)` clean/dirty counts reproduces the single-threaded
-    /// tally exactly, independent of the order the workers' indices are merged.
-    fn merge(&mut self, other: Self) {
-        self.total = self.total.saturating_add(other.total);
-        for (set, by_commit) in other.sets {
-            let dest = self.sets.entry(set).or_default();
-            for (topo_index, counts) in by_commit {
-                let entry = dest.entry(topo_index).or_insert_with(|| CommitCounts {
-                    commit: counts.commit.clone(),
-                    clean: 0,
-                    dirty: 0,
-                });
-                entry.clean = entry.clean.saturating_add(counts.clean);
-                entry.dirty = entry.dirty.saturating_add(counts.dirty);
-            }
-        }
-    }
-
-    /// Total runs admitted across every set.
-    pub(crate) fn total(&self) -> usize {
-        self.total
-    }
-
-    /// Whether no run entered the selection.
-    pub(crate) fn is_empty(&self) -> bool {
-        self.total == 0
-    }
-
-    /// Runs admitted in `set`.
-    pub(crate) fn runs_in_set(&self, set: &DiscriminantSet) -> usize {
-        self.sets.get(set).map_or(0, |by_commit| {
-            by_commit
-                .values()
-                .map(|counts| counts.clean.saturating_add(counts.dirty))
-                .sum()
-        })
-    }
-
-    /// Each set with at least one run, paired with its per-commit tallies in
-    /// first-parent topological order (oldest first).
-    pub(crate) fn sets(
-        &self,
-    ) -> impl Iterator<Item = (&DiscriminantSet, &BTreeMap<usize, CommitCounts>)> {
-        self.sets.iter()
-    }
-
-    /// The oldest and newest commit that contributed a run, by first-parent
-    /// topological position, as `(first, last)` full SHAs. `None` when no run was
-    /// admitted. The report header uses it to state the span of analyzed history.
-    // The oldest/newest tie-break is unobservable, so its comparison mutants are
-    // equivalent: a first-parent topological position denotes exactly one commit, so
-    // every set records the same commit ID at a given position and `<` vs `<=` (or `>` vs
-    // `>=`) only ever chooses between identical strings.
-    #[cfg_attr(test, mutants::skip)]
-    pub(crate) fn commit_span(&self) -> Option<(&str, &str)> {
-        // A given first-parent position maps to exactly one commit, so every set records
-        // the same commit under it: the span is simply the commit at the lowest position
-        // and the one at the highest. Reading those extremes with `min_by_key`/`max_by_key`
-        // keeps the ordering in the standard library rather than a hand-rolled comparison.
-        let first = self
-            .sets
-            .values()
-            .flat_map(|by_commit| by_commit.iter())
-            .min_by_key(|entry| *entry.0)?
-            .1
-            .commit
-            .as_str();
-        let last = self
-            .sets
-            .values()
-            .flat_map(|by_commit| by_commit.iter())
-            .max_by_key(|entry| *entry.0)?
-            .1
-            .commit
-            .as_str();
-        Some((first, last))
-    }
-}
-
-/// The data an analysis (or listing) draws on, plus the bookkeeping needed to
-/// explain an empty outcome and warn about ephemeral data.
-pub(crate) struct SelectedDataSet {
-    /// The reconstructed series for the in-window runs, built with the caller's
-    /// series filter and ordered by git topology. Pre-blessing: the caller applies
-    /// blessings (history mode) or leaves them unapplied (branch/tip, listings).
-    pub(crate) series: Vec<Series>,
-    /// Compact per-set, per-commit run tallies, standing in for a retained copy of
-    /// every loaded object (which a large history cannot afford to keep resident).
-    pub(crate) run_index: RunIndex,
-    /// How many facet-matching candidates existed before topology filtering.
-    pub(crate) candidate_count: usize,
-    /// Why candidates were excluded, for the empty-history hint.
-    pub(crate) tally: ExclusionTally,
-    /// Whether a dirty run was admitted solely by the base-branch dirty-tree
-    /// exception (triggers the ephemeral-data warning).
-    pub(crate) included_dirty_base_exception: bool,
-    /// The target ref the timeline was resolved against (for diagnostics).
-    pub(crate) target_ref: String,
-    /// Subject line of each in-history commit that has one, so `examine` can label
-    /// each data point with what its commit changed. A commit absent here has an
-    /// empty subject; only `examine` reads this.
-    pub(crate) commit_subjects: HashMap<String, String>,
-    /// The full commit ID of the analyzed tip commit (the resolved `--context`/HEAD),
-    /// carried into the report so it names the exact commit the findings describe.
-    pub(crate) tip_commit: String,
-    /// Whether the working tree carried uncommitted changes when the analysis ran;
-    /// the report annotates the tip `+ uncommitted changes` when set.
-    pub(crate) tip_dirty: bool,
-    /// The resolved analysis mode (auto-detected from topology, or overridden).
-    pub(crate) mode: AnalysisMode,
-    /// First-parent topological index of the merge-base, used by branch mode to
-    /// split base-side history from the branch's own commits.
-    pub(crate) merge_base_index: Option<usize>,
-    /// Blessings recorded on in-window commits, grouped by discriminant set. Each
-    /// entry pairs the blessed commit's first-parent topological index and its
-    /// committer date (from topology, for the report anchor) with the record;
-    /// history-mode re-baselining picks, per series, the latest matching blessing.
-    /// Empty in branch and tip modes (they ignore blessings).
-    pub(crate) blessings: HashMap<DiscriminantSet, Vec<BlessingPlacement>>,
-}
-
-/// Resolves one facet's raw command-line values into a [`FacetFilter`].
-///
-/// The case-insensitive `all` keyword (anywhere in the list) is an explicit
-/// synonym for no filter. An empty list auto-detects: the current-machine value
-/// when one is supplied (`auto`), else no filter (engine has no host default).
-fn resolve_facet(values: &[String], auto: Option<&str>) -> FacetFilter {
-    if values.iter().any(|value| value.eq_ignore_ascii_case("all")) {
-        return FacetFilter::All;
-    }
-    match NonEmpty::from_vec(values.to_vec()) {
-        Some(values) => FacetFilter::Explicit(values),
-        None => auto.map_or(FacetFilter::All, |value| {
-            FacetFilter::Auto(value.to_owned())
-        }),
-    }
-}
-
-/// Resolves every command-line facet into a [`DiscriminantSetQuery`] filter, validating that any
-/// explicit `--engine` values name a known engine.
-///
-/// `auto` supplies the current-machine defaults for the triple and machine-key
-/// facets when those are omitted. Passing `None` resolves omitted facets to no
-/// filter instead — used by the `discriminants` catalog listing, which is a
-/// discovery view over all stored partitions rather than the current machine's.
-pub(crate) fn resolve_facets(
-    selection: &Selection<'_>,
-    auto: Option<&AutoFacets>,
-) -> Result<DiscriminantSetQuery, RunError> {
-    let engine = resolve_facet(selection.engine, None);
-    if let FacetFilter::Explicit(values) = &engine {
-        for value in values.iter() {
-            parse_engine(Some(value))?;
-        }
-    }
-    Ok(DiscriminantSetQuery {
-        engine,
-        target_triple: resolve_facet(selection.target_triple, auto.map(|a| a.triple.as_str())),
-        machine_key: resolve_facet(selection.machine_key, auto.map(|a| a.machine_key.as_str())),
-    })
-}
-
-/// Lists the stored objects under the project's partition and keeps the ones whose
-/// discriminant set matches the facet filters. Shared by the topology-aware
-/// selection and the discriminant listing (which needs no repository).
-pub(crate) async fn facet_filtered_candidates<S: Storage>(
-    storage: &S,
-    project_id: &str,
-    facets: &DiscriminantSetQuery,
-    reporter: &dyn Reporter,
-) -> Result<Vec<(String, StorageKey)>, RunError> {
-    // The listing prefix must use the same sanitized project segment that
-    // `DiscriminantSet` writes its storage keys under. A project id containing a
-    // character that sanitizes (a space, `/`, a non-ASCII letter, ...) is stored
-    // mangled, so listing under the raw id would silently find an empty history.
-    let project = sanitize_segment(project_id);
-    let prefix = project_objects_prefix(project_id);
-
-    reporter.if_enabled(|notes| {
-        notes.note(&format!(
-            "project id: {project_id} (storage segment: {project})"
-        ));
-        notes.note(&format!("listing stored objects under prefix {prefix}"));
-        notes.note(&format!("facet filters: {}", describe_facets(facets)));
-    });
-
-    let list_started = Instant::now();
-    let keys = storage.list(&prefix).await.map_err(RunError::Storage)?;
-    reporter.timing("storage.list(prefix) round-trip", list_started.elapsed());
-    reporter.note_with(|| format!("storage returned {}", count_noun(keys.len(), "object key")));
-
-    let mut candidates: Vec<(String, StorageKey)> = Vec::new();
-    for key in keys {
-        if !key.ends_with(".json") {
-            reporter.note_with(|| format!("skipping {key}: not a .json object"));
-            continue;
-        }
-        let Some(parsed) = parse_key(&key) else {
-            reporter.note_with(|| {
-                format!("skipping {key}: not a recognized {STORAGE_VERSION} storage key")
-            });
-            continue;
-        };
-        if !facets.matches(&parsed.set) {
-            reporter.note_with(|| {
-                format!(
-                    "skipping {key}: discriminant {} does not match the facet filters",
-                    parsed.set
-                )
-            });
-            continue;
-        }
-        candidates.push((key, parsed));
-    }
-    reporter.note_with(|| {
-        format!(
-            "{} match the facet filters",
-            count_noun(candidates.len(), "object")
-        )
-    });
-    Ok(candidates)
-}
-
-/// How many stored objects to fetch concurrently while loading a data set.
-///
-/// `analyze`/`list` load every in-selection object before reconstructing the
-/// series. Each [`Storage::get`] is a round-trip — a *network* round-trip against
-/// the Azure Blob backend — so fetching them one at a time makes the load a sum of
-/// latencies, which dominates wall time at scale. Overlapping a bounded number of
-/// fetches turns that sum into roughly its maximum, cutting the per-mode load
-/// floor (critical for the remote backend, where thousands of sequential
-/// round-trips would otherwise stretch the local floor into minutes). The bound
-/// sits near the knee of the throughput curve: enough fetches are in flight to
-/// saturate the network path and hide per-object latency, while staying below the
-/// point where extra concurrency merely subdivides the fixed path bandwidth among
-/// more requests and lengthens each one's latency without lifting throughput.
-const LOAD_CONCURRENCY: usize = 128;
-
-/// Fetches and deserializes the given stored objects with bounded concurrency.
-///
-/// `parse` turns one object's raw bytes into the parsed value `T` (it owns the
-/// UTF-8 decoding so the per-type error wording stays exact). The fetches overlap
-/// up to [`LOAD_CONCURRENCY`] at a time and therefore complete out of order, so
-/// the caller must re-sort the results (by storage key) to keep diagnostics and
-/// the loaded order deterministic. The whole operation stays single-threaded and
-/// `!Send`, so it runs unchanged under the Miri-driven `block_on` tests.
-async fn load_objects_concurrently<S, T, F>(
-    storage: &S,
-    keys: Vec<(String, StorageKey)>,
-    parse: F,
-) -> Result<Vec<(String, StorageKey, T)>, RunError>
-where
-    S: Storage,
-    F: Fn(&str, Vec<u8>) -> Result<T, RunError>,
-{
-    let parse = &parse;
-    futures::stream::iter(keys)
-        .map(move |(key, parsed)| fetch_one(storage, key, parsed, parse))
-        .buffer_unordered(LOAD_CONCURRENCY)
-        .try_collect::<Vec<_>>()
-        .await
-}
-
-/// Fetches and deserializes a single stored object. Factored out of
-/// [`load_objects_concurrently`] so the stream closure stays a plain `FnMut`
-/// returning this future (rather than a closure wrapping an `async` block).
-async fn fetch_one<S, T, F>(
-    storage: &S,
-    key: String,
-    parsed: StorageKey,
-    parse: &F,
-) -> Result<(String, StorageKey, T), RunError>
-where
-    S: Storage,
-    F: Fn(&str, Vec<u8>) -> Result<T, RunError>,
-{
-    let bytes = storage.get(&key).await.map_err(RunError::Storage)?;
-    let value = parse(&key, bytes)?;
-    Ok((key, parsed, value))
-}
-
-/// One worker's (or the merged) folded contribution: the series builder it folded
-/// its object chunk into, the run tally it recorded, and the per-object admission
-/// flags for the key-ordered verbose notes.
-///
-/// Returned from each spawned fold task and merged on the main thread; the merged
-/// value carries the same shape so the caller treats one worker and many uniformly.
-struct WorkerFold {
-    /// Series points folded from this worker's objects (compact; the parsed runs
-    /// are dropped inside the worker as each is folded).
-    builder: SeriesBuilder,
-    /// Per-set, per-commit run tally for this worker's objects.
-    run_index: RunIndex,
-    /// `(storage key, admitted-by-dirty-base-exception)` per folded object, for the
-    /// key-ordered verbose notes the caller emits once every worker has folded.
-    admitted: Vec<(String, bool)>,
-}
-
-/// Loads, parses, and **folds** the in-selection survivors across CPU cores.
-///
-/// `ranked` is the storage-key-sorted survivor list, each carrying its
-/// storage-key ordinal (`rank`). It is split into [`worker_count`] balanced
-/// contiguous chunks ([`balanced_chunk_sizes`]); one spawned task per chunk
-/// fetches, decompresses, parses, and folds its slice into its *own*
-/// [`SeriesBuilder`] / [`RunIndex`], dropping each parsed run the moment its
-/// compact points are extracted. The main thread awaits the chunks in spawn order
-/// and merges their builders, run tallies, and admission lists into one
-/// [`WorkerFold`]; a single [`SeriesBuilder::finish`] then sorts the combined
-/// series. The merge is associative and the final sort is global, so the result is
-/// identical to a single-threaded fold in storage-key order.
-///
-/// Folding inside each worker (rather than buffering every chunk's parsed runs and
-/// folding serially on the main thread) parallelizes the fold as well as the parse:
-/// a worker drops each parsed run the moment its compact points are extracted, so it
-/// never buffers its chunk's parsed runs. This does **not** lower peak memory — at
-/// merge time every worker's finished builder is resident at once alongside the
-/// growing combined builder (~2x the compact point output transiently), which measured
-/// ~5% above the buffered-parallel variant; the win it buys is the ~10% faster wall
-/// time and better core use from moving the fold off the main thread. The decompress +
-/// JSON parse — the CPU-dominated cost — is spread across the runtime's worker threads.
-/// See the "Architecture" section of `docs/DESIGN.md` and the load section of
-/// `docs/analyze.md`.
-async fn fold_runs_chunked<S>(
-    storage: &S,
-    spawner: &Spawner,
-    ranked: Vec<(usize, String, StorageKey)>,
-    order: &Arc<HashMap<String, usize>>,
-    dirty_base_exception: &Arc<HashMap<String, bool>>,
-    prefixes: Arc<[BenchmarkIdPrefix]>,
-) -> Result<WorkerFold, RunError>
-where
-    S: Storage + Clone + 'static,
-{
-    let total = ranked.len();
-    let mut combined = WorkerFold {
-        builder: SeriesBuilder::with_prefixes(Arc::clone(&prefixes)),
-        run_index: RunIndex::new(),
-        admitted: Vec::with_capacity(total),
-    };
-    if total == 0 {
-        return Ok(combined);
-    }
-    let workers = worker_count(total);
-
-    let mut items = ranked.into_iter();
-    let mut handles = Vec::with_capacity(workers);
-    for chunk_len in balanced_chunk_sizes(total, workers) {
-        let chunk: Vec<(usize, String, StorageKey)> = items.by_ref().take(chunk_len).collect();
-        let storage = storage.clone();
-        let order = Arc::clone(order);
-        let dirty_base_exception = Arc::clone(dirty_base_exception);
-        let prefixes = Arc::clone(&prefixes);
-        handles.push(spawner.spawn(async move {
-            let mut builder = SeriesBuilder::with_prefixes(prefixes);
-            let mut run_index = RunIndex::new();
-            let mut admitted: Vec<(String, bool)> = Vec::with_capacity(chunk.len());
-            for (rank, key, parsed) in chunk {
-                let bytes = storage.get(&key).await.map_err(RunError::Storage)?;
-                let text = str::from_utf8(&bytes).map_err(|error| RunError::Analyze {
-                    message: format!("stored object {key} is not valid UTF-8: {error}"),
-                })?;
-                let run = RunPoints::from_json(text).map_err(|error| RunError::Analyze {
-                    message: format!("stored object {key} is not a valid result set: {error}"),
-                })?;
-                let topo_index = order
-                    .get(&parsed.commit)
-                    .copied()
-                    .expect("phase 1 admitted only commits on the analyzed history");
-                let dirty = parsed.is_dirty();
-                let is_exception = dirty
-                    && dirty_base_exception
-                        .get(parsed.commit.as_str())
-                        .copied()
-                        .unwrap_or(false);
-                run_index.record(&parsed.set, topo_index, &parsed.commit, dirty);
-                builder.push(
-                    &parsed.set,
-                    topo_index,
-                    dirty,
-                    ordinal_of(rank),
-                    &parsed.commit,
-                    &run,
-                );
-                admitted.push((key, is_exception));
-                // `run` is dropped here; only the extracted (compact) points are kept.
-            }
-            Ok::<WorkerFold, RunError>(WorkerFold {
-                builder,
-                run_index,
-                admitted,
-            })
-        }));
-    }
-
-    for handle in handles {
-        let fold = handle.await?;
-        combined.builder.merge(fold.builder);
-        combined.run_index.merge(fold.run_index);
-        combined.admitted.extend(fold.admitted);
-    }
-    Ok(combined)
-}
-
-/// Resolves the git topology, selects the comparable commits, and loads the
-/// in-selection objects into a [`SelectedDataSet`]. Requires a repository: the
-/// timeline is reconstructed from git history, not from stored timestamps.
-#[expect(
-    clippy::too_many_arguments,
-    reason = "mirrors the analyze selection pipeline, which threads the same injected ports"
-)]
-pub(crate) async fn select_dataset<G, S>(
-    git: &G,
-    storage: &S,
-    project_id: &str,
-    config: &Config,
-    selection: &Selection<'_>,
-    filter: SeriesFilter<'_>,
-    auto: &AutoFacets,
-    now: Timestamp,
-    reporter: &dyn Reporter,
-    spawner: &Spawner,
-) -> Result<SelectedDataSet, RunError>
-where
-    G: GitHistory,
-    S: Storage + Clone + 'static,
-{
-    let facets = resolve_facets(selection, Some(auto))?;
-    let listing_started = Instant::now();
-    let candidates = facet_filtered_candidates(storage, project_id, &facets, reporter).await?;
-    reporter.timing(
-        "candidate listing + facet filter (includes storage.list)",
-        listing_started.elapsed(),
-    );
-
-    // Separate blessing sidecars from run objects: they share the partition prefix
-    // but carry a different payload and are loaded into their own map rather than
-    // the series.
-    let (candidates, bless_candidates): (Vec<_>, Vec<_>) = candidates
-        .into_iter()
-        .partition(|(_, parsed)| !parsed.is_bless());
-    if !bless_candidates.is_empty() {
-        reporter.note_with(|| {
-            format!(
-                "{} of those are blessing sidecars",
-                count_noun(bless_candidates.len(), "object")
-            )
-        });
-    }
-
-    let topology_started = Instant::now();
-    let ResolvedHistory {
-        target_ref,
-        tip_commit,
-        tip_dirty,
-        order,
-        commit_times,
-        commit_subjects,
-        admit_dirty,
-        dirty_base_exception,
-        merge_base_index,
-        tip_is_merge_base,
-    } = resolve_history(
-        git,
-        config,
-        selection,
-        DirtyTipPolicy::WhenWorkingTreeDirty,
-        reporter,
-    )
-    .await?;
-    reporter.timing(
-        "git topology resolution (resolve_history)",
-        topology_started.elapsed(),
-    );
-
-    // The topology lookups the parallel fold needs — the commit -> first-parent
-    // index map and the base-branch dirty-tree exceptions — are shared read-only
-    // across every worker, so wrap them in `Arc` once and hand each worker a cheap
-    // clone. Downstream main-thread reads still go through `Deref`.
-    let order = Arc::new(order);
-    let dirty_base_exception = Arc::new(dirty_base_exception);
-
-    // Mode auto-detection keys off the *recorded data set*, not the on-disk
-    // repository state. The branch view exists to compare a feature branch's runs
-    // against its base, so it only applies when there is actually feature-branch
-    // data: commits past the merge-base, or a dirty run recorded on top of the
-    // base tip. A dirty working tree with no dirty run stored on the tip carries no
-    // such data, so it stays the long-range history view.
-    let dirty_tip_run_present = candidates.iter().any(|(_, parsed)| {
-        parsed.is_dirty()
-            && dirty_base_exception
-                .get(parsed.commit.as_str())
-                .copied()
-                .unwrap_or(false)
-    });
-
-    // The mode steers the analysis and the default `--since`. An explicit `--mode`
-    // overrides the auto-detection.
-    let mode =
-        match selection.mode_override {
-            Some(mode) => {
-                reporter.note_with(|| {
-                    format!(
-                        "analysis mode: {} (set explicitly via --mode, overriding auto-detection)",
-                        mode.as_str()
-                    )
-                });
-                mode
-            }
-            None => {
-                let mode = auto_mode(tip_is_merge_base, dirty_tip_run_present);
-                reporter.note_with(|| format!(
-                "analysis mode: {} (auto-detected because the target tip {} its own merge-base \
-                 with the base branch and {} recorded on top of it; the on-disk working-tree \
-                 state is deliberately not consulted here)",
-                mode.as_str(),
-                if tip_is_merge_base { "is" } else { "is not" },
-                if dirty_tip_run_present {
-                    "a dirty run is"
-                } else {
-                    "no dirty run is"
-                },
-            ));
-                mode
-            }
-        };
-    let since = resolve_since(selection.since, mode, now)?;
-    reporter.note_with(|| {
-        format!(
-            "since cutoff: {} ({})",
-            since.map_or_else(|| "none".to_owned(), |since| since.to_string()),
-            since_cutoff_reason(selection.since.is_some(), mode)
-        )
-    });
-    let until = parse_until(selection.until, now)?;
-    reporter.note_with(|| {
-        format!(
-            "until cutoff: {}",
-            until.map_or_else(|| "none".to_owned(), |until| until.to_string())
-        )
-    });
-
-    // Tally why candidates do not enter the analysis, so a `0 runs` outcome can
-    // explain itself (via `--verbose` per object, and via a summary hint when
-    // candidates existed but none were admitted).
-    let candidate_count = candidates.len();
-    let mut excluded_outside_history = 0_usize;
-    let mut excluded_dirty_base = 0_usize;
-    let mut excluded_since = 0_usize;
-    let mut excluded_until = 0_usize;
-
-    // Phase 1 — key-only filtering, in candidate order. Every exclusion that does
-    // not need the object's payload runs here, before anything is fetched, so an
-    // excluded candidate never costs a round-trip: history membership, base-side
-    // dirty admission, and the `--since`/`--until` window (decided from each
-    // commit's committer time, which git reports with the topology).
-    let phase1_started = Instant::now();
-    let mut to_fetch: Vec<(String, StorageKey)> = Vec::new();
-    for (key, parsed) in candidates {
-        if !order.contains_key(&parsed.commit) {
-            excluded_outside_history = excluded_outside_history.saturating_add(1);
-            reporter.note_with(|| {
-                format!(
-                    "excluding {key}: commit {} is not on {target_ref}'s analyzed history",
-                    parsed.commit
-                )
-            });
-            continue;
-        }
-        if parsed.is_dirty()
-            && !admit_dirty
-                .get(parsed.commit.as_str())
-                .copied()
-                .unwrap_or(false)
-        {
-            excluded_dirty_base = excluded_dirty_base.saturating_add(1);
-            reporter.note_with(|| {
-                format!(
-                    "excluding {key}: dirty snapshot on a base-side commit ({} \
-                     only admits clean runs); dirty runs count only on the target side",
-                    parsed.commit
-                )
-            });
-            continue;
-        }
-        match window_excludes(commit_times.get(&parsed.commit).copied(), since, until) {
-            Some(WindowEdge::Since) => {
-                excluded_since = excluded_since.saturating_add(1);
-                reporter.note_with(|| {
-                    format!(
-                        "excluding {key}: commit {} is before the --since cutoff",
-                        parsed.commit
-                    )
-                });
-                continue;
-            }
-            Some(WindowEdge::Until) => {
-                excluded_until = excluded_until.saturating_add(1);
-                reporter.note_with(|| {
-                    format!(
-                        "excluding {key}: commit {} is after the --until cutoff",
-                        parsed.commit
-                    )
-                });
-                continue;
-            }
-            None => {}
-        }
-        to_fetch.push((key, parsed));
-    }
-    reporter.timing(
-        "phase 1 — key-only candidate filtering (no fetches)",
-        phase1_started.elapsed(),
-    );
-
-    // Phase 2/3 — fetch the survivors and fold each into the series. The fetch +
-    // decompress + JSON parse (the CPU-dominated cost) is spread across the
-    // runtime's worker threads: the storage-key-sorted survivors are split into
-    // balanced contiguous chunks, and one spawned task fetches, parses, *and folds*
-    // each chunk into its own series builder — dropping each parsed run the instant
-    // its compact points are extracted. The main thread then merges the per-worker
-    // builders, run tallies, and admission lists into one. Each object's ordinal —
-    // the final point tie-break — is its rank in storage-key order, assigned up
-    // front, so the single `builder.finish()` sort reproduces the in-order result.
-    // The per-object verbose notes are collected during the fold and emitted in
-    // storage-key order afterwards, so the diagnostics stay byte-identical to a
-    // deterministic in-order pass.
-    //
-    // Folding inside each worker keeps only the compact per-worker points resident
-    // between fetch and merge — never the whole parsed data set — so the parallel
-    // parse does not buy its throughput with the full-buffer memory peak.
-    to_fetch.sort_by(|left, right| left.0.cmp(&right.0));
-
-    let fetch_fold_started = Instant::now();
-    let ranked: Vec<(usize, String, StorageKey)> = to_fetch
-        .into_iter()
-        .enumerate()
-        .map(|(rank, (key, parsed))| (rank, key, parsed))
-        .collect();
-    let prefixes: Arc<[BenchmarkIdPrefix]> = Arc::from(filter.prefixes);
-    let WorkerFold {
-        builder,
-        run_index,
-        mut admitted,
-    } = fold_runs_chunked(
-        storage,
-        spawner,
-        ranked,
-        &order,
-        &dirty_base_exception,
-        prefixes,
-    )
-    .await?;
-    // Whether at least one dirty run was admitted solely by the base-branch
-    // dirty-tree exception, so the report can warn that it is ephemeral.
-    let included_dirty_base_exception = admitted.iter().any(|(_, is_exception)| *is_exception);
-    reporter.timing(
-        "phase 2/3 — chunked parallel fetch + parse + per-worker fold, then merge",
-        fetch_fold_started.elapsed(),
-    );
-
-    // Emit the per-object verbose notes in storage-key order — the deterministic
-    // order objects were previously admitted in — then the summary.
-    admitted.sort_by(|left, right| left.0.cmp(&right.0));
-    for (key, is_exception) in &admitted {
-        if *is_exception {
-            reporter.note_with(|| {
-                format!(
-                    "including {key}: dirty snapshot on the base-branch tip, admitted \
-                     because the working tree is dirty (ephemeral — see the warning)"
-                )
-            });
-        } else {
-            reporter.note_with(|| format!("including {key}"));
-        }
-    }
-    let finish_started = Instant::now();
-    let series = builder.finish();
-    reporter.timing(
-        "series build finalization (builder.finish: assemble + serial point sort)",
-        finish_started.elapsed(),
-    );
-    reporter.note_with(|| {
-        format!(
-            "{} entered the analysis ({excluded_outside_history} outside history, \
-         {excluded_dirty_base} dirty-on-base, {excluded_since} before --since, \
-         {excluded_until} after --until)",
-            count_noun(run_index.total(), "object")
-        )
-    });
-
-    // Load the blessing sidecars on in-window commits into a per-set map. A
-    // blessing on a commit outside the analyzed history (or that fails to parse) is
-    // irrelevant and skipped. Branch and tip modes ignore blessings entirely, so
-    // only history mode pays the load.
-    let mut blessings: HashMap<DiscriminantSet, Vec<BlessingPlacement>> = HashMap::new();
-    if mode == AnalysisMode::History {
-        let blessing_started = Instant::now();
-        // Phase 1 — key-only filtering: drop blessings whose commit is not on the
-        // analyzed history before fetching, in candidate order.
-        let mut to_fetch: Vec<(String, StorageKey)> = Vec::new();
-        for (key, parsed) in bless_candidates {
-            if order.contains_key(&parsed.commit) {
-                to_fetch.push((key, parsed));
-            } else {
-                reporter.note_with(|| {
-                    format!(
-                        "skipping blessing {key}: commit {} is not on {target_ref}'s analyzed \
-                         history",
-                        parsed.commit
-                    )
-                });
-            }
-        }
-        // Phase 2 — fetch and deserialize concurrently, then restore storage-key
-        // order (`buffer_unordered` completes out of order).
-        let mut fetched = load_objects_concurrently(storage, to_fetch, |key, bytes| {
-            let text = String::from_utf8(bytes).map_err(|error| RunError::Analyze {
-                message: format!("stored blessing {key} is not valid UTF-8: {error}"),
-            })?;
-            BlessingRecord::from_json(&text).map_err(|error| RunError::Analyze {
-                message: format!("stored blessing {key} is not a valid blessing record: {error}"),
-            })
-        })
-        .await?;
-        fetched.sort_by(|left, right| left.0.cmp(&right.0));
-        // Phase 3 — record each blessing against its commit's topological index
-        // and committer date (resolved from topology, for the report anchor).
-        for (key, parsed, record) in fetched {
-            let topo_index = order
-                .get(&parsed.commit)
-                .copied()
-                .expect("phase 1 admitted only blessings whose commit is on the analyzed history");
-            let commit_time = commit_times.get(&parsed.commit).copied();
-            reporter.note_with(|| {
-                format!(
-                    "loaded blessing {key} ({} accepted at {})",
-                    count_noun(record.prefixes.len(), "prefix filter"),
-                    parsed.commit
-                )
-            });
-            blessings.entry(parsed.set.clone()).or_default().push((
-                topo_index,
-                commit_time,
-                record,
-            ));
-        }
-        reporter.timing(
-            "blessing sidecar load (history mode: filter + fetch + parse)",
-            blessing_started.elapsed(),
-        );
-    }
-
-    Ok(SelectedDataSet {
-        series,
-        run_index,
-        candidate_count,
-        tally: ExclusionTally {
-            outside_history: excluded_outside_history,
-            dirty_base: excluded_dirty_base,
-            since: excluded_since,
-            until: excluded_until,
-        },
-        included_dirty_base_exception,
-        target_ref,
-        commit_subjects,
-        tip_commit,
-        tip_dirty,
-        mode,
-        merge_base_index,
-        blessings,
-    })
-}
-
-/// Narrows a storage-key rank to the series point ordinal width.
-///
-/// The ordinal is a pure tie-break, so the (practically impossible) overflow past
-/// `u32::MAX` distinct in-window objects merely lets the last ordinals collide —
-/// the affected points then keep their stable fold order rather than panicking.
-#[expect(
-    clippy::cast_possible_truncation,
-    reason = "saturating: ordinals only tie-break, and >4 billion in-window objects never occur"
-)]
-fn ordinal_of(rank: usize) -> u32 {
-    rank.min(u32::MAX as usize) as u32
-}
-
-/// Auto-detects the analysis mode from the resolved topology and recorded data.
-///
-/// An official base-branch view — the target's tip *is* its own merge-base (or no
-/// base is known) and no dirty run is recorded on that tip — is
-/// [`AnalysisMode::History`]. Anything else (commits past the merge-base, or a
-/// dirty run actually recorded on top of the base tip) is treated as an unnamed
-/// feature branch: [`AnalysisMode::Branch`]. The detection looks only at the data
-/// set, never at the on-disk working-tree state, so a dirty checkout with no dirty
-/// run stored on the tip still analyzes as history. [`AnalysisMode::Tip`] is never
-/// auto-selected.
-fn auto_mode(tip_is_merge_base: bool, dirty_tip_run_present: bool) -> AnalysisMode {
-    if tip_is_merge_base && !dirty_tip_run_present {
-        AnalysisMode::History
-    } else {
-        AnalysisMode::Branch
-    }
-}
-
-/// Explains, for a verbose note, why the resolved `--since` cutoff is what it is.
-fn since_cutoff_reason(explicit_since: bool, mode: AnalysisMode) -> &'static str {
-    if explicit_since {
-        "from the --since option"
-    } else if mode == AnalysisMode::History {
-        "history-mode default six-month look-back"
-    } else {
-        "no default look-back window outside history mode"
-    }
-}
-
-/// Resolves the effective `--since` cutoff: an explicit value always wins;
-/// otherwise history mode applies a default look-back so a scheduled trend watch
-/// does not silently widen as history accumulates, while branch and tip modes have
-/// no default (a feature branch's whole history is in scope).
-fn resolve_since(
-    value: Option<&str>,
-    mode: AnalysisMode,
-    now: Timestamp,
-) -> Result<Option<Timestamp>, RunError> {
-    if value.is_some() {
-        return parse_since(value, now);
-    }
-    if mode == AnalysisMode::History {
-        return default_history_since(now).map(Some);
-    }
-    Ok(None)
-}
-
-/// Default history-mode look-back window: six months before `now`.
-const HISTORY_DEFAULT_LOOKBACK_MONTHS: i32 = 6;
-
-/// The instant [`HISTORY_DEFAULT_LOOKBACK_MONTHS`] before `now`, anchored with
-/// calendar-correct zoned arithmetic (months have no fixed length).
-fn default_history_since(now: Timestamp) -> Result<Timestamp, RunError> {
-    now.to_zoned(TimeZone::UTC)
-        .checked_sub(Span::new().months(HISTORY_DEFAULT_LOOKBACK_MONTHS))
-        .map(|zoned| zoned.timestamp())
-        .map_err(|error| RunError::Analyze {
-            message: format!("default --since window is out of the representable range: {error}"),
-        })
-}
-
-/// Parses the `--mode` option into an explicit [`AnalysisMode`] override.
-///
-/// `auto` (the default when omitted) resolves to `None` so the mode is detected
-/// from topology; `history`, `branch`, and `tip` force that mode.
-fn parse_mode(value: Option<&str>) -> Result<Option<AnalysisMode>, RunError> {
-    match value {
-        None | Some("auto") => Ok(None),
-        Some(name) => AnalysisMode::from_name(name)
-            .map(Some)
-            .ok_or_else(|| RunError::Analyze {
-                message: format!(
-                    "unknown analysis mode {name:?}; expected auto, history, branch, or tip"
-                ),
-            }),
-    }
-}
-
-/// How the base-branch dirty-tip exception is gated.
-///
-/// On a feature branch the target-side commits admit dirty runs unconditionally;
-/// this policy only governs the *base* branch's tip. `analyze`/`list` admit a
-/// base-tip dirty run only when the working tree is currently dirty (the
-/// "evaluating the tool / accidentally on the base branch" case); `prune` admits
-/// base-tip dirty runs regardless of the current working-tree state.
-#[derive(Clone, Copy)]
-pub(crate) enum DirtyTipPolicy {
-    /// Admit a base-side tip's dirty runs only when the working tree is dirty now.
-    WhenWorkingTreeDirty,
-    /// Always treat a base-side tip as admitting dirty runs.
-    Always,
-}
-
-/// The git topology a selection resolves to: the target ref it was resolved
-/// against, the first-parent position of each selected commit, and the per-commit
-/// dirty-admission flags. All maps use owned commit IDs so the borrowed
-/// `selected` set can drop before the caller's load loop.
-pub(crate) struct ResolvedHistory {
-    /// The target ref the timeline was resolved against (for diagnostics).
-    pub(crate) target_ref: String,
-    /// The full commit ID the target ref resolved to — the analyzed tip commit, carried
-    /// into the report so it names the exact commit the findings describe.
-    pub(crate) tip_commit: String,
-    /// Whether the working tree carried uncommitted changes when the topology was
-    /// resolved. Probed only under [`DirtyTipPolicy::WhenWorkingTreeDirty`] (and
-    /// not under `--no-dirty`); `false` otherwise. The report annotates the tip
-    /// `+ uncommitted changes` when set.
-    pub(crate) tip_dirty: bool,
-    /// First-parent position of each selected commit, for series ordering. An
-    /// object whose commit is absent is outside the analyzed history.
-    pub(crate) order: HashMap<String, usize>,
-    /// Committer timestamp of each first-parent commit, for deciding the
-    /// `--since`/`--until` window from topology before any object is fetched. A
-    /// commit absent here has an unknown time and is treated as in-window.
-    pub(crate) commit_times: HashMap<String, Timestamp>,
-    /// Subject line of each first-parent commit that has one, for labeling
-    /// `examine`'s per-commit data points. A commit absent here has an empty
-    /// subject; only `examine` reads this.
-    pub(crate) commit_subjects: HashMap<String, String>,
-    /// Whether each selected commit admits dirty (uncommitted-tree) snapshots.
-    pub(crate) admit_dirty: HashMap<String, bool>,
-    /// Whether a commit's dirty runs are admitted *only* by the base-branch
-    /// dirty-tree exception, which triggers the ephemeral-data warning.
-    pub(crate) dirty_base_exception: HashMap<String, bool>,
-    /// First-parent topological index of the merge-base, used by branch mode to
-    /// split base-side history from the branch's own commits. `None` when no base
-    /// is known or the merge-base is off the analyzed chain.
-    pub(crate) merge_base_index: Option<usize>,
-    /// Whether the target's tip *is* its own merge-base with the base (or no base
-    /// is known): the signal that this is an official base-branch view.
-    pub(crate) tip_is_merge_base: bool,
-}
-
-/// Resolves the git topology for a selection: the target ref's first-parent
-/// ancestry, the merge-base with the base ref, and the per-commit dirty-admission
-/// flags. Requires a repository — an unresolvable target ref is an error rather
-/// than an empty success.
-pub(crate) async fn resolve_history<G>(
-    git: &G,
-    config: &Config,
-    selection: &Selection<'_>,
-    policy: DirtyTipPolicy,
-    reporter: &dyn Reporter,
-) -> Result<ResolvedHistory, RunError>
-where
-    G: GitHistory,
-{
-    // Resolving the timeline requires a repository: the topology comes from git
-    // history, not from stored timestamps. An unresolvable target ref means there
-    // is no repository here (or the branch does not exist), which is an error.
-    let target_ref = selection.context.unwrap_or("HEAD");
-    let Some(target_commit_id) = git.resolve(target_ref).await.map_err(RunError::Io)? else {
-        return Err(RunError::Analyze {
-            message: format!(
-                "this command requires a git repository: could not resolve {target_ref:?}. \
-                 Run inside a repository (or pass --repo / --context)."
-            ),
-        });
-    };
-
-    let base_commit_id = resolve_base_ref(git, config, selection.base).await?;
-    let first_parent_started = Instant::now();
-    let first_parent = git
-        .first_parent(&target_commit_id)
-        .await
-        .map_err(RunError::Io)?;
-    reporter.timing(
-        "git.first_parent ancestry walk (target's first-parent line)",
-        first_parent_started.elapsed(),
-    );
-    // Split the first-parent ancestry into the commit ID timeline (for commit selection
-    // and the merge-base lookup) and a commit ID -> committer-time map (for the window).
-    let commit_count = first_parent.len();
-    let mut ancestry: Vec<String> = Vec::with_capacity(commit_count);
-    let mut commit_times: HashMap<String, Timestamp> = HashMap::new();
-    let mut commit_subjects: HashMap<String, String> = HashMap::new();
-    for commit in first_parent {
-        if let Some(time) = commit.committer_time {
-            commit_times.insert(commit.commit_id.clone(), time);
-        }
-        if !commit.subject.is_empty() {
-            commit_subjects.insert(commit.commit_id.clone(), commit.subject);
-        }
-        ancestry.push(commit.commit_id);
-    }
-    let merge_base = match &base_commit_id {
-        Some(base) => git
-            .merge_base(&target_commit_id, base)
-            .await
-            .map_err(RunError::Io)?,
-        None => None,
-    };
-
-    reporter.if_enabled(|notes| {
-        notes.note(&format!(
-            "target ref {target_ref} resolves to {target_commit_id}; {} on its first-parent line",
-            count_noun(commit_count, "commit")
-        ));
-        notes.note(&format!(
-            "base ref resolves to {}; merge-base with target is {}",
-            base_commit_id.as_deref().unwrap_or("<none>"),
-            merge_base.as_deref().unwrap_or("<none>")
-        ));
-    });
-
-    // The base-branch dirty-tip exception: `analyze`/`list` admit a base-side tip's
-    // dirty runs only when the working tree is currently dirty (`--no-dirty` skips
-    // both the probe and the exception); `prune` admits them unconditionally so it
-    // can remove them regardless of the present working-tree state. The probe result
-    // is reused for the report's tip annotation, so `analyze` never runs it twice.
-    let working_tree_dirty = match policy {
-        DirtyTipPolicy::WhenWorkingTreeDirty if !selection.no_dirty => {
-            git.is_dirty().await.map_err(RunError::Io)?
-        }
-        _ => false,
-    };
-    let dirty_tip_exception = match policy {
-        DirtyTipPolicy::Always => !selection.no_dirty,
-        DirtyTipPolicy::WhenWorkingTreeDirty => {
-            if working_tree_dirty {
-                reporter.note_with(|| {
-                    "working tree is dirty: dirty snapshots on a base-side tip will be admitted"
-                        .to_owned()
-                });
-            }
-            working_tree_dirty
-        }
-    };
-
-    let selected = select_commits(
-        &ancestry,
-        merge_base.as_deref(),
-        !selection.no_dirty,
-        dirty_tip_exception,
-    );
-    let order: HashMap<String, usize> = selected
-        .iter()
-        .enumerate()
-        .map(|(index, one)| (one.commit.clone(), index))
-        .collect();
-    let admit_dirty: HashMap<String, bool> = selected
-        .iter()
-        .map(|one| (one.commit.clone(), one.admit_dirty))
-        .collect();
-    let dirty_base_exception: HashMap<String, bool> = selected
-        .iter()
-        .map(|one| (one.commit.clone(), one.dirty_base_exception))
-        .collect();
-
-    // The merge-base's topological position (when it is on the analyzed chain)
-    // splits base-side history from the branch's own commits in branch mode.
-    let merge_base_index = merge_base
-        .as_deref()
-        .and_then(|base| order.get(base).copied());
-    // The target's tip is its own merge-base (or no base is known) exactly when
-    // this is an official base-branch view rather than a feature branch.
-    let tip_is_merge_base = merge_base
-        .as_deref()
-        .is_none_or(|base| base == target_commit_id);
-
-    Ok(ResolvedHistory {
-        target_ref: target_ref.to_owned(),
-        tip_commit: target_commit_id,
-        tip_dirty: working_tree_dirty,
-        order,
-        commit_times,
-        commit_subjects,
-        admit_dirty,
-        dirty_base_exception,
-        merge_base_index,
-        tip_is_merge_base,
-    })
-}
-
-/// The ephemeral-data warning appended when a dirty base-branch-tip run is admitted.
-pub(crate) fn dirty_base_exception_warning() -> String {
-    "Warning: analysis included dirty runs (with uncommitted changes) on top of the \
-     base branch. These may be excluded from future analysis. Switch to a new branch \
-     to persist benchmark history of your changes."
-        .to_owned()
-}
-
-/// Resolves the base ref the target's history is split against, returning its
-/// commit ID or `None` when no base can be determined.
-///
-/// Precedence: an explicit `--base` (an error if it does not resolve), then the
-/// configured `project.default_branch`, then the repository's detected default
-/// branch (`origin/HEAD`, else `main`/`master`).
-pub(crate) async fn resolve_base_ref<G: GitHistory>(
-    git: &G,
-    config: &Config,
-    base: Option<&str>,
-) -> Result<Option<String>, RunError> {
-    if let Some(base) = base {
-        return git
-            .resolve(base)
-            .await
-            .map_err(RunError::Io)?
-            .map(Some)
-            .ok_or_else(|| RunError::Analyze {
-                message: format!("could not resolve --base {base:?}"),
-            });
-    }
-    if let Some(default) = config.project.default_branch.as_deref()
-        && let Some(resolved) = git.resolve(default).await.map_err(RunError::Io)?
-    {
-        return Ok(Some(resolved));
-    }
-    if let Some(name) = git.default_branch().await.map_err(RunError::Io)?
-        && let Some(resolved) = git.resolve(&name).await.map_err(RunError::Io)?
-    {
-        return Ok(Some(resolved));
-    }
-    Ok(None)
-}
-
-/// Resolves the *display name* of the base ref (without resolving it to a commit ID),
-/// for diagnostics such as the `prune --prune-base` guard.
-///
-/// Mirrors [`resolve_base_ref`]'s precedence: an explicit `--base`, then the
-/// configured `project.default_branch` (only when it resolves), then the
-/// repository's detected default branch. Returns `None` when no base can be named.
-pub(crate) async fn resolve_base_name<G: GitHistory>(
-    git: &G,
-    config: &Config,
-    base: Option<&str>,
-) -> Result<Option<String>, RunError> {
-    if let Some(base) = base {
-        return Ok(Some(base.to_owned()));
-    }
-    if let Some(default) = config.project.default_branch.as_deref()
-        && git.resolve(default).await.map_err(RunError::Io)?.is_some()
-    {
-        return Ok(Some(default.to_owned()));
-    }
-    git.default_branch().await.map_err(RunError::Io)
-}
-
-/// A human-readable summary of the active facet filters, for `--verbose` notes.
-fn describe_facets(facets: &DiscriminantSetQuery) -> String {
-    let parts = [
-        ("engine", &facets.engine),
-        ("target_triple", &facets.target_triple),
-        ("machine_key", &facets.machine_key),
-    ]
-    .into_iter()
-    .filter_map(|(label, filter)| describe_filter(filter).map(|value| format!("{label}={value}")))
-    .collect::<Vec<_>>();
-    if parts.is_empty() {
-        "none".to_owned()
-    } else {
-        parts.join(", ")
-    }
-}
-
-/// Renders one facet filter, or `None` when it imposes no constraint.
-fn describe_filter(filter: &FacetFilter) -> Option<String> {
-    match filter {
-        FacetFilter::All => None,
-        FacetFilter::Auto(value) => Some(format!("{value} (auto-detected)")),
-        FacetFilter::Explicit(values) => Some(values.iter().cloned().collect::<Vec<_>>().join("|")),
-    }
-}
-
-/// How many facet-matching candidates were excluded, by reason.
-#[derive(Clone, Copy, Debug)]
-pub(crate) struct ExclusionTally {
-    /// Commit is not on the analyzed first-parent history.
-    outside_history: usize,
-    /// Dirty snapshot on a base-side commit (clean-only).
-    dirty_base: usize,
-    /// Effective time is before the `--since` cutoff.
-    since: usize,
-    /// Effective time is after the `--until` cutoff.
-    until: usize,
-}
-
-/// Builds a diagnostic hint for the case where stored runs matched the facet
-/// filters but none entered the analysis, so the empty outcome explains itself.
-///
-/// Returns `None` when at least one run was loaded, or when there were no
-/// candidates at all (a genuinely empty history needs no special explanation).
-pub(crate) fn empty_history_hint(
-    loaded_is_empty: bool,
-    candidate_count: usize,
-    target_ref: &str,
-    tally: ExclusionTally,
-) -> Option<String> {
-    if !loaded_is_empty || candidate_count == 0 {
-        return None;
-    }
-
-    let mut lines = vec![format!(
-        "Found {} for this project, but none entered the analysis:",
-        count_noun(candidate_count, "stored run")
-    )];
-    if tally.dirty_base > 0 {
-        lines.push(format!(
-            "  - {} on base-branch commits — only clean runs count on the base \
-             branch. Commit your working tree (including the configuration file) and re-run, \
-             or analyze a feature branch with --context.",
-            count_noun(tally.dirty_base, "dirty (uncommitted-tree) snapshot")
-        ));
-    }
-    if tally.outside_history > 0 {
-        lines.push(format!(
-            "  - {} on commits outside {target_ref}'s analyzed history — check out the \
-             branch they were recorded on, or pass --context.",
-            count_noun(tally.outside_history, "run")
-        ));
-    }
-    if tally.since > 0 {
-        lines.push(format!(
-            "  - {} older than the --since cutoff.",
-            count_noun(tally.since, "run")
-        ));
-    }
-    if tally.until > 0 {
-        lines.push(format!(
-            "  - {} newer than the --until cutoff.",
-            count_noun(tally.until, "run")
-        ));
-    }
-    lines.push("Re-run with --verbose for a per-object explanation.".to_owned());
-    Some(lines.join("\n"))
-}
-
-/// Parses an `--engine` facet value into an [`Engine`], if set.
-fn parse_engine(name: Option<&str>) -> Result<Option<Engine>, RunError> {
-    match name {
-        None => Ok(None),
-        Some(name) => Engine::from_name(name)
-            .map(Some)
-            .ok_or_else(|| RunError::Analyze {
-                message: format!(
-                    "unknown engine {name:?}; expected one of: criterion, callgrind, \
-                     alloc_tracker, all_the_time"
-                ),
-            }),
-    }
-}
-
-/// Parses the `--since` option into an absolute lower-bound instant, if set.
-///
-/// See [`parse_instant`] for the accepted input forms. `now` anchors a relative
-/// duration.
-pub(crate) fn parse_since(
-    value: Option<&str>,
-    now: Timestamp,
-) -> Result<Option<Timestamp>, RunError> {
-    parse_instant(value, "--since", now)
-}
-
-/// Parses the `--until` option into an absolute upper-bound instant, if set.
-///
-/// See [`parse_instant`] for the accepted input forms. `now` anchors a relative
-/// duration.
-pub(crate) fn parse_until(
-    value: Option<&str>,
-    now: Timestamp,
-) -> Result<Option<Timestamp>, RunError> {
-    parse_instant(value, "--until", now)
-}
-
-/// Which edge of the `--since`/`--until` window a commit falls outside of.
-#[derive(Clone, Copy, Debug, Eq, PartialEq)]
-pub(crate) enum WindowEdge {
-    /// The commit predates the `--since` lower bound.
-    Since,
-    /// The commit postdates the `--until` upper bound.
-    Until,
-}
-
-/// Decides whether a commit's `committer_time` places it outside the
-/// `--since`/`--until` window, reporting which edge it fell off.
-///
-/// `--since` is an inclusive-on-or-after lower bound and `--until` an
-/// inclusive-on-or-before upper bound, matching the object-timestamp comparison
-/// the analysis used before topology carried the time. A commit with an unknown
-/// time (`None`) is never excluded — git always reports a committer date for a
-/// real commit, so this only guards a degenerate input conservatively (by
-/// keeping the object and letting the rest of the pipeline judge it).
-pub(crate) fn window_excludes(
-    committer_time: Option<Timestamp>,
-    since: Option<Timestamp>,
-    until: Option<Timestamp>,
-) -> Option<WindowEdge> {
-    let time = committer_time?;
-    if since.is_some_and(|since| time < since) {
-        return Some(WindowEdge::Since);
-    }
-    if until.is_some_and(|until| time > until) {
-        return Some(WindowEdge::Until);
-    }
-    None
-}
-
-/// Parses a time-cutoff option into an absolute instant, if set.
-///
-/// Three input forms are accepted, tried in order:
-///
-/// * an RFC 3339 timestamp (`2024-01-01T00:00:00Z`),
-/// * a bare `YYYY-MM-DD` date, interpreted at UTC midnight, and
-/// * a relative duration in jiff's friendly or ISO 8601 form (`5 months`,
-///   `5 months ago`, `P6M`, `2w`), interpreted as *that far in the past* —
-///   resolved against `now` via calendar-correct zoned arithmetic.
-///
-/// `now` is the analysis anchor sourced from the injected clock (see
-/// [`resolve_now`]), so the relative form is deterministic under a frozen clock
-/// rather than reading the wall clock afresh.
-///
-/// The relative form is normalized through [`Span::abs`] before subtracting, so a
-/// duration written with the friendly `ago` suffix (which jiff parses as a
-/// *negative* span) still means "this far back" rather than flipping into the
-/// future. A cutoff in the future is never a sensible bound, so both `5 months`
-/// and `5 months ago` resolve to the same past instant.
-fn parse_instant(
-    value: Option<&str>,
-    flag: &str,
-    now: Timestamp,
-) -> Result<Option<Timestamp>, RunError> {
-    let Some(value) = value else {
-        return Ok(None);
-    };
-    let value = value.trim();
-    if let Ok(timestamp) = value.parse::<Timestamp>() {
-        return Ok(Some(timestamp));
-    }
-    if let Ok(date) = value.parse::<Date>() {
-        // UTC has no DST transitions, so civil midnight always maps to an instant.
-        let zoned = date
-            .to_zoned(TimeZone::UTC)
-            .expect("UTC midnight is always a valid instant");
-        return Ok(Some(zoned.timestamp()));
-    }
-    if let Ok(span) = value.parse::<Span>() {
-        return Ok(Some(instant_before(span, flag, now)?));
-    }
-    Err(RunError::Analyze {
-        message: format!(
-            "invalid {flag} value {value:?}; expected an RFC 3339 timestamp, a YYYY-MM-DD \
-             date, or a relative duration such as \"6 months\" or \"30 days ago\""
-        ),
-    })
-}
-
-/// Resolves a relative [`Span`] to the instant that far before `now`, treating the
-/// span's magnitude as a look-back regardless of its sign (see [`parse_instant`]).
-///
-/// Calendar units (months, years) have no fixed length, so the subtraction is
-/// anchored to `now`'s UTC zoned datetime rather than to a bare duration.
-fn instant_before(span: Span, flag: &str, now: Timestamp) -> Result<Timestamp, RunError> {
-    now.to_zoned(TimeZone::UTC)
-        .checked_sub(span.abs())
-        .map(|zoned| zoned.timestamp())
-        .map_err(|error| RunError::Analyze {
-            message: format!("{flag} duration is out of the representable range: {error}"),
-        })
-}
-
 #[cfg(test)]
 #[cfg_attr(coverage_nightly, coverage(off))]
 mod tests {
@@ -1856,8 +310,8 @@ mod tests {
     use crate::config::{Config, parse_config};
     use crate::git_history::FakeGitHistory;
     use crate::model::{
-        BenchmarkId, BenchmarkIdPrefix, BenchmarkResult, EnvironmentInfo, GitInfo, Metric,
-        MetricKind, Run, RunContext, ToolchainInfo,
+        BenchmarkId, BenchmarkIdPrefix, BenchmarkResult, BlessingRecord, EnvironmentInfo, GitInfo,
+        Metric, MetricKind, Run, RunContext, ToolchainInfo, sanitize_segment,
     };
     use crate::output::MemoryOutputWriter;
     use crate::report::RecordingReporter;
@@ -2144,166 +598,11 @@ mod tests {
     }
 
     #[test]
-    fn auto_mode_uses_tip_topology_and_recorded_dirty_runs_not_repo_state() {
-        // A base branch whose tip is its own merge-base with no dirty run recorded
-        // on the tip is history mode.
-        assert_eq!(auto_mode(true, false), AnalysisMode::History);
-        // Commits past the merge-base, or a dirty run actually recorded on top of
-        // the tip, make it a (possibly unnamed) feature branch.
-        assert_eq!(auto_mode(false, false), AnalysisMode::Branch);
-        assert_eq!(auto_mode(true, true), AnalysisMode::Branch);
-        assert_eq!(auto_mode(false, true), AnalysisMode::Branch);
-    }
-
-    #[test]
-    fn since_cutoff_reason_explains_each_source() {
-        assert_eq!(
-            since_cutoff_reason(true, AnalysisMode::History),
-            "from the --since option"
-        );
-        assert_eq!(
-            since_cutoff_reason(true, AnalysisMode::Branch),
-            "from the --since option"
-        );
-        assert_eq!(
-            since_cutoff_reason(false, AnalysisMode::History),
-            "history-mode default six-month look-back"
-        );
-        assert_eq!(
-            since_cutoff_reason(false, AnalysisMode::Branch),
-            "no default look-back window outside history mode"
-        );
-    }
-
-    #[test]
     fn should_colorize_only_in_an_interactive_terminal_without_no_color() {
         assert!(should_colorize(true, false), "terminal, NO_COLOR unset");
         assert!(!should_colorize(false, false), "not a terminal");
         assert!(!should_colorize(true, true), "NO_COLOR set");
         assert!(!should_colorize(false, true), "neither");
-    }
-
-    #[test]
-    fn describe_facets_joins_set_facets_and_reports_none_when_empty() {
-        let empty = DiscriminantSetQuery {
-            engine: FacetFilter::All,
-            target_triple: FacetFilter::All,
-            machine_key: FacetFilter::All,
-        };
-        assert_eq!(describe_facets(&empty), "none");
-
-        let full = DiscriminantSetQuery {
-            engine: FacetFilter::Explicit(nonempty!["criterion".to_owned()]),
-            target_triple: FacetFilter::Auto("x86_64-pc-windows-msvc".to_owned()),
-            machine_key: FacetFilter::Explicit(nonempty!["abcd".to_owned()]),
-        };
-        assert_eq!(
-            describe_facets(&full),
-            "engine=criterion, target_triple=x86_64-pc-windows-msvc (auto-detected), \
-             machine_key=abcd"
-        );
-    }
-
-    #[test]
-    fn empty_history_hint_explains_only_when_runs_were_excluded() {
-        let no_exclusions = ExclusionTally {
-            outside_history: 0,
-            dirty_base: 0,
-            since: 0,
-            until: 0,
-        };
-        // No candidates at all → a genuinely empty history needs no hint.
-        assert_eq!(empty_history_hint(true, 0, "master", no_exclusions), None);
-        // Runs were actually loaded → no hint.
-        assert_eq!(empty_history_hint(false, 3, "master", no_exclusions), None);
-
-        let tally = ExclusionTally {
-            outside_history: 2,
-            dirty_base: 1,
-            since: 4,
-            until: 0,
-        };
-        let hint = empty_history_hint(true, 7, "master", tally).unwrap();
-        assert!(hint.contains("7 stored runs"), "{hint}");
-        assert!(
-            hint.contains("1 dirty (uncommitted-tree) snapshot"),
-            "{hint}"
-        );
-        assert!(hint.contains("2 runs on commits outside master"), "{hint}");
-        assert!(
-            hint.contains("4 runs older than the --since cutoff"),
-            "{hint}"
-        );
-        assert!(hint.contains("--verbose"), "{hint}");
-
-        // A zero reason omits its line entirely (each `> 0` guard is exercised in
-        // both directions): only the dirty reason is present here.
-        let dirty_only = ExclusionTally {
-            outside_history: 0,
-            dirty_base: 3,
-            since: 0,
-            until: 0,
-        };
-        let hint = empty_history_hint(true, 3, "master", dirty_only).unwrap();
-        assert!(hint.contains("dirty (uncommitted-tree)"), "{hint}");
-        assert!(!hint.contains("outside"), "{hint}");
-        assert!(!hint.contains("--since cutoff"), "{hint}");
-        assert!(!hint.contains("--until cutoff"), "{hint}");
-
-        // Only the outside-history reason is present here (dirty omitted).
-        let outside_only = ExclusionTally {
-            outside_history: 2,
-            dirty_base: 0,
-            since: 0,
-            until: 0,
-        };
-        let hint = empty_history_hint(true, 2, "master", outside_only).unwrap();
-        assert!(hint.contains("outside master"), "{hint}");
-        assert!(!hint.contains("dirty (uncommitted-tree)"), "{hint}");
-        assert!(!hint.contains("--since cutoff"), "{hint}");
-
-        // Only the until reason is present here.
-        let until_only = ExclusionTally {
-            outside_history: 0,
-            dirty_base: 0,
-            since: 0,
-            until: 5,
-        };
-        let hint = empty_history_hint(true, 5, "master", until_only).unwrap();
-        assert!(
-            hint.contains("5 runs newer than the --until cutoff"),
-            "{hint}"
-        );
-        assert!(!hint.contains("dirty (uncommitted-tree)"), "{hint}");
-        assert!(!hint.contains("--since cutoff"), "{hint}");
-    }
-
-    #[test]
-    fn parse_mode_resolves_auto_to_none_and_rejects_unknown() {
-        assert_eq!(parse_mode(None).unwrap(), None);
-        assert_eq!(parse_mode(Some("auto")).unwrap(), None);
-        assert_eq!(
-            parse_mode(Some("history")).unwrap(),
-            Some(AnalysisMode::History)
-        );
-        assert_eq!(
-            parse_mode(Some("branch")).unwrap(),
-            Some(AnalysisMode::Branch)
-        );
-        assert_eq!(parse_mode(Some("tip")).unwrap(), Some(AnalysisMode::Tip));
-        let error = parse_mode(Some("weekly")).unwrap_err();
-        assert!(
-            error.to_string().contains("unknown analysis mode"),
-            "{error}"
-        );
-    }
-
-    #[test]
-    fn parse_engine_resolves_none_known_and_rejects_unknown() {
-        assert!(parse_engine(None).unwrap().is_none());
-        assert!(parse_engine(Some("callgrind")).unwrap().is_some());
-        let error = parse_engine(Some("nonsuch")).unwrap_err();
-        assert!(error.to_string().contains("unknown"), "{error}");
     }
 
     #[test]
@@ -2333,66 +632,6 @@ mod tests {
             "{:?}",
             reporter.notes()
         );
-    }
-
-    #[test]
-    fn resolve_since_applies_a_default_only_in_history_mode() {
-        let now = "2024-06-01T00:00:00Z".parse::<Timestamp>().unwrap();
-        // History mode without an explicit value falls back to the six-month window.
-        let history = resolve_since(None, AnalysisMode::History, now)
-            .unwrap()
-            .unwrap();
-        assert_eq!(
-            history,
-            "2023-12-01T00:00:00Z".parse::<Timestamp>().unwrap()
-        );
-        // Branch and tip modes have no default cutoff.
-        assert_eq!(
-            resolve_since(None, AnalysisMode::Branch, now).unwrap(),
-            None
-        );
-        assert_eq!(resolve_since(None, AnalysisMode::Tip, now).unwrap(), None);
-        // An explicit value always wins, even in branch mode.
-        let explicit = resolve_since(Some("2024-01-01"), AnalysisMode::Branch, now)
-            .unwrap()
-            .unwrap();
-        assert_eq!(
-            explicit,
-            "2024-01-01T00:00:00Z".parse::<Timestamp>().unwrap()
-        );
-    }
-
-    #[test]
-    fn default_history_since_subtracts_six_calendar_months() {
-        let now = "2024-03-31T00:00:00Z".parse::<Timestamp>().unwrap();
-        // Calendar arithmetic, not a fixed number of days: six months before
-        // 2024-03-31 is 2023-09-30 (September has 30 days).
-        let cutoff = default_history_since(now).unwrap();
-        assert_eq!(cutoff, "2023-09-30T00:00:00Z".parse::<Timestamp>().unwrap());
-    }
-
-    #[test]
-    fn window_excludes_decides_each_edge_from_committer_time() {
-        let since = Some(ts(10));
-        let until = Some(ts(20));
-        // Before the lower bound, after the upper bound, and inside the window.
-        assert_eq!(
-            window_excludes(Some(ts(5)), since, until),
-            Some(WindowEdge::Since)
-        );
-        assert_eq!(
-            window_excludes(Some(ts(25)), since, until),
-            Some(WindowEdge::Until)
-        );
-        assert_eq!(window_excludes(Some(ts(15)), since, until), None);
-        // Both bounds are inclusive: a commit exactly on an edge stays in-window.
-        assert_eq!(window_excludes(Some(ts(10)), since, until), None);
-        assert_eq!(window_excludes(Some(ts(20)), since, until), None);
-        // An open bound never excludes on that side.
-        assert_eq!(window_excludes(Some(ts(0)), None, until), None);
-        assert_eq!(window_excludes(Some(ts(99)), since, None), None);
-        // An unknown committer time is never excluded, even with both bounds set.
-        assert_eq!(window_excludes(None, since, until), None);
     }
 
     /// Runs `analyze_with` and unwraps the rendered report and regression count.
@@ -3580,207 +1819,11 @@ mod tests {
     }
 
     #[test]
-    fn since_accepts_timestamp_and_date() {
-        let now: Timestamp = "2024-06-01T00:00:00Z".parse().unwrap();
-        assert_eq!(
-            parse_since(Some("2024-01-01T00:00:00Z"), now).unwrap(),
-            Some("2024-01-01T00:00:00Z".parse().unwrap())
-        );
-        assert_eq!(
-            parse_since(Some("2024-01-01"), now).unwrap(),
-            Some("2024-01-01T00:00:00Z".parse().unwrap())
-        );
-        assert_eq!(parse_since(None, now).unwrap(), None);
-    }
-
-    #[test]
-    fn since_accepts_relative_durations_as_look_back() {
-        // A friendly duration resolves to an instant in the past, and the `ago`
-        // suffix (which jiff parses as a negative span) means the same look-back
-        // rather than flipping into the future. Anchored to a fixed `now`, both
-        // spellings land on exactly the same instant — five calendar months
-        // before 2024-06-01 is 2024-01-01.
-        let now: Timestamp = "2024-06-01T00:00:00Z".parse().unwrap();
-        let expected: Timestamp = "2024-01-01T00:00:00Z".parse().unwrap();
-        let plain = parse_since(Some("5 months"), now).unwrap().unwrap();
-        let with_ago = parse_since(Some("5 months ago"), now).unwrap().unwrap();
-        assert_eq!(plain, expected, "5 months before 2024-06-01 is 2024-01-01");
-        assert_eq!(
-            with_ago, expected,
-            "the `ago` suffix looks back the same amount"
-        );
-        assert!(plain < now, "a look-back must be in the past");
-    }
-
-    #[test]
-    fn since_accepts_iso_and_week_durations() {
-        let now: Timestamp = "2024-06-01T00:00:00Z".parse().unwrap();
-        for input in ["P6M", "2w", "30 days", "-P1Y"] {
-            let cutoff = parse_since(Some(input), now).unwrap().unwrap();
-            assert!(cutoff < now, "{input:?} must resolve to the past");
-        }
-    }
-
-    #[test]
-    fn since_rejects_garbage() {
-        let now: Timestamp = "2024-06-01T00:00:00Z".parse().unwrap();
-        let error = parse_since(Some("not-a-date"), now).unwrap_err();
-        assert!(matches!(error, RunError::Analyze { .. }), "{error:?}");
-    }
-
-    #[test]
-    fn run_index_counts_runs_and_reports_emptiness() {
-        let set = DiscriminantSet {
-            engine: "criterion".to_owned(),
-            target_triple: "x86_64-unknown-linux-gnu".to_owned(),
-            machine_key: "synthetic".to_owned(),
-        };
-        let mut index = RunIndex::new();
-        assert!(index.is_empty(), "a fresh index admits no runs");
-        assert_eq!(index.total(), 0);
-        assert_eq!(index.runs_in_set(&set), 0);
-
-        index.record(&set, 0, "c0", false);
-        index.record(&set, 0, "c0", true);
-        index.record(&set, 1, "c1", false);
-
-        assert!(
-            !index.is_empty(),
-            "recording even one run makes the index non-empty"
-        );
-        assert_eq!(index.total(), 3, "every recorded run is counted once");
-        assert_eq!(
-            index.runs_in_set(&set),
-            3,
-            "clean and dirty runs both count toward the set tally"
-        );
-    }
-
-    #[test]
-    fn run_index_merge_sums_totals_and_per_commit_counts() {
-        // Each worker records its disjoint object chunk into its own index; merging
-        // the per-worker indices must reproduce the single-threaded tally exactly,
-        // summing the totals and the per-(set, commit) clean/dirty counts.
-        let set = DiscriminantSet {
-            engine: "criterion".to_owned(),
-            target_triple: "x86_64-unknown-linux-gnu".to_owned(),
-            machine_key: "synthetic".to_owned(),
-        };
-        let other_set = DiscriminantSet {
-            engine: "callgrind".to_owned(),
-            target_triple: "x86_64-unknown-linux-gnu".to_owned(),
-            machine_key: "synthetic".to_owned(),
-        };
-
-        // The reference index folds every run in one pass.
-        let mut reference = RunIndex::new();
-        reference.record(&set, 0, "c0", false);
-        reference.record(&set, 0, "c0", true);
-        reference.record(&set, 1, "c1", false);
-        reference.record(&other_set, 0, "c0", false);
-
-        // Two workers split the same runs; c0/set is touched by both so the merge has
-        // to sum the per-commit counts rather than overwrite them.
-        let mut first = RunIndex::new();
-        first.record(&set, 0, "c0", false);
-        first.record(&other_set, 0, "c0", false);
-        let mut second = RunIndex::new();
-        second.record(&set, 0, "c0", true);
-        second.record(&set, 1, "c1", false);
-
-        first.merge(second);
-
-        assert_eq!(first.total(), reference.total());
-        assert_eq!(first.runs_in_set(&set), reference.runs_in_set(&set));
-        assert_eq!(
-            first.runs_in_set(&other_set),
-            reference.runs_in_set(&other_set)
-        );
-
-        let summarize = |index: &RunIndex| {
-            index
-                .sets()
-                .map(|(set, by_commit)| {
-                    let counts: Vec<(usize, String, usize, usize)> = by_commit
-                        .iter()
-                        .map(|(topo, counts)| {
-                            (*topo, counts.commit.clone(), counts.clean, counts.dirty)
-                        })
-                        .collect();
-                    (set.clone(), counts)
-                })
-                .collect::<Vec<_>>()
-        };
-        assert_eq!(summarize(&first), summarize(&reference));
-    }
-
-    #[test]
-    fn commit_span_spans_the_oldest_and_newest_analyzed_commit() {
-        let set = DiscriminantSet {
-            engine: "criterion".to_owned(),
-            target_triple: "x86_64-unknown-linux-gnu".to_owned(),
-            machine_key: "synthetic".to_owned(),
-        };
-        let other_set = DiscriminantSet {
-            engine: "callgrind".to_owned(),
-            target_triple: "x86_64-unknown-linux-gnu".to_owned(),
-            machine_key: "synthetic".to_owned(),
-        };
-
-        let mut index = RunIndex::new();
-        assert_eq!(index.commit_span(), None, "an empty index spans nothing");
-
-        // Record out of topological order and across two sets: the span must key off
-        // the first-parent position, not the insertion order, and must consider every
-        // set so the header covers the whole analyzed history.
-        index.record(&set, 2, "c2", false);
-        index.record(&set, 0, "c0", false);
-        index.record(&other_set, 1, "c1", false);
-
-        assert_eq!(
-            index.commit_span(),
-            Some(("c0", "c2")),
-            "the span runs from the lowest to the highest topological position"
-        );
-    }
-
-    #[test]
-    fn commit_span_collapses_to_a_single_commit() {
-        let set = DiscriminantSet {
-            engine: "criterion".to_owned(),
-            target_triple: "x86_64-unknown-linux-gnu".to_owned(),
-            machine_key: "synthetic".to_owned(),
-        };
-        let mut index = RunIndex::new();
-        index.record(&set, 0, "solo", false);
-        index.record(&set, 0, "solo", true);
-
-        assert_eq!(
-            index.commit_span(),
-            Some(("solo", "solo")),
-            "a single analyzed commit is both ends of the span"
-        );
-    }
-
-    #[test]
     fn resolve_now_reads_the_injected_clock() {
         // The analyze family sources its wall-clock anchor through an injectable
         // `tick::Clock`; a frozen clock must surface its own instant verbatim rather than
         // any default minted independently of the clock.
         let anchor = ts(1_700_000_000);
         assert_eq!(resolve_now(Some(Clock::new_frozen_at(anchor))), anchor);
-    }
-
-    #[test]
-    fn ordinal_of_passes_small_ranks_through_unchanged() {
-        // The ordinal is the storage-key rank narrowed to the point's `u32` width;
-        // realistic ranks pass through verbatim so the series tie-break stays in
-        // key order.
-        assert_eq!(ordinal_of(0), 0);
-        assert_eq!(ordinal_of(7), 7);
-        assert_eq!(
-            ordinal_of(usize::try_from(u32::MAX).expect("u32 fits in usize")),
-            u32::MAX
-        );
     }
 }
